Use `define` for constants, and use more warnings.
Using `define` to silence a lint found by `clang`
(`-Wgnu-folding-constant`). Apparently, I was using
VLAs by using a `const int` to declare an array size
and "folding it to a constant" is a GNU extension (gcc).
Clang doesn't like that, and I think I can get by with
just defining them as constants anyway. We'll see.

Enabling more warnings flags for the build.
Also a `clang` target was added, for testing purposes.