`fore_arg` and `fore` are working dynamically.
...still lots to do, and lots of dead code to remove when
it's done. Translating those two to a `back_arg`/`back`, or
`style_arg`/`style` is no problem. The problem is joining arbitrary
values to build one long string.

The plan may be to use a variadic ColorArg-based function, and
the `fore` macro just builds one of those. So,

```c
C(fore(RED), "Test ", back("lightblue"), "this", ".")
```

Expands to:
```c
ColorArgs_to_str(
    ColorArg_from_value(fore(RED)),
    ColorArg_from_value("Test"),
    ...,
)
```
Which means ColorArg needs a `TYPE_TEXT`, that simply returns the
text when converted to str.

Still thinking on this.