Arg-parsing with `getopt`, rainbow tool for fg/bg.
The colr tool now accepts positional or flag arguments for fore, back,
and style. Arguments can be passed in any order with the flags.

`"rainbow"` is a working fore/back color for the colr tool. Work needs
to be done to `Colr()`, to make it recognize `"rainbow"` as a color
name (impossible in `colr()`). This way, all the colr tool has to do is
forward the user's fore/back args to `Colr()`, and users of `ColrC`
can take advantage of the feature also.