Add `sanitize` target for colrc and tests.
The tests are built with it by default.

The colr tool is tested in sanitized mode by running:
```bash
./test/run_tests.sh -A
```

It already caught one bug in the *_is_ptr() functions (already fixed).

`colrc` should not be built with the `sanitize` options in release mode,
because it is way too slow, and uses way too much memory. Only use it
for tests.