Add `colr_str_get_codes`, and more esc-code funcs.
`colr_str_get_codes`: Allocates a list of escape-codes found in a
string (`NULL` if no escape-codes are present). The strings are
allocated and copied, and the pointer-list is allocated.

`colr_free_str_list`: Frees a list like the one created by
`colr_str_get_codes`.

`colr_str_code_cnt`: Count the number of escape-codes in a string.

`colr_str_code_len`: Count the number of escape-code chars in a string.