Remove basic/ext/RGB_str_static macros. Not-sound.
libasan was complaining all over the place because I was addressing
stack that didn't belong to me. The fore/back/style_str_static macros
don't seem to be affected, and they're the only ones that are needed
anyway. The others were just the foundation/experiment to build on.

It turns out I didn't need them anyway, so good riddance.