Only store filename in info, and store line num.
This will enable recreating a Line() from a tuple of
(linenum, text) (actually just `Line(*tup)` should do it).

The entire file path isn't important, but the parent dir might be.
'test/test_this.c' may be useful, or 'examples/colr_example.c', but
'colr/colr.c' may not be. I haven't decided yet. The other advantage
of storing only the filename is testing if the name is in a certain
file. As in, `info[name]['files'].get('colr.c', None).get('count',
0)`.