unused.py: Load/save results, Names hold the info.
It's a little easier to test against a Name if it already has all of it's
info in the class.

More work needs to be done to clean things up, like doing
`name.info['total']` or `name.info['files']` is a little ugly.
It could be `name.total` or `name.files`.

It was a quick coversion, with JSON input/output in mind. The
JSON/info-gathering functions will need to be rewritten with the
new style in mind. `check_file()` and `check_files()` are even
more out of sync now. One deals with dicts and the other deals with
a list of `Name`s. It needs to be separated like that to keep from
scanning the files over and over for every name, but the end result
could be, and will be cleaner when I'm done.