100% function coverage, 99% line coverage.
The other 1% are edge cases dealing with `TYPE_NONE` and `TYPE_INVALID`
stuff. They can, and will, be tested.

There are certain allocation-failure-branches that are not covered,
and they are excluded from the report. I don't know how to test those
right now, but I can say that those lines are very simple (1-2 lines)
and they usually return `NULL`, which is checked for in every function
that receives a pointer to anything.