`alloca` vs VLA docs, warnings, and cppcheck.
Warnings/notes were added in the docs about the use of `alloca` in
`back_str_static` and `fore_str_static`. These are not needed for normal
use of ColrC, and they were GNU-specific to begin with, but links to
the GNU docs and a Stack Overflow question were added.

The cppcheck suppressions files were updated to ignore false-positives.
CppCheck is wonky sometimes, and 99% of the warnings are bogus or not
important to the test binary.

The one warning that was real (check return value of `sprintf`) was
satisfied with an `assert(written_len == expected_len)`. At least
debug builds will be checked for that one situation. It's a function
that's mainly used for debugging/testing anyway.